requirements.txt
black==19.10b0
slackclient==2.6.0
falcon==2.0.0
gunicorn==20.0.4